feat(core,tests,security): modernize ctop with 80.7% test coverage, security hardening, and defect resolutions
- Rebrand and migrate module path to github.com/edsilegx/ctop
- Implement comprehensive 14-group test catalog achieving 80.7% statement coverage
- Fix 10 identified defects including widget deadlocks, 292-yr uptime overflow, and G115 integer overflows
- Ensure all containers (running, paused, stopped, created) are visible by default
- Upgrade runtime dependencies (runc v1.5.1, selinux v1.15.1, go-dockerclient v1.13.2)
- Pass Gosec, Semgrep, ShellCheck, GolangCI-Lint, and TruffleHog audits with 0 issues
- Enable hardened -trimpath -buildmode=pie compilation on Linux and Windows
- Update comprehensive README.md, ARCHITECTURE.md, TESTING.md, CHANGELOG.md, and install.sh

+++ models/main.go
@@ -1,12 +1,17 @@
+// Package models defines core telemetry, container metadata, and log data structures.
+// Objective: Provide lightweight, standardized domain types for container metrics, log lines, and metadata.
+// Data Flow: Connector Telemetry Stream -> Models (Metrics, Log, Meta) -> Container -> UI Widgets.
 package models
 
 import "time"
 
+// Log represents a single timestamped container log entry.
 type Log struct {
 	Timestamp time.Time
 	Message   string
 }
 
+// Meta holds arbitrary key-value metadata strings associated with a container.
 type Meta map[string]string
 
 // NewMeta returns an initialized Meta map.
@@ -31,7 +36,7 @@ func (m Meta) Get(k string) string {
 }
 
 type Metrics struct {
-	NCpus        uint8
+	NCpus        int
 	CPUUtil      int
 	NetTx        int64
 	NetRx        int64
